♻️ refactor: 파일 및 폴더 구조 리팩토링 #6
