+++ src/app/onboarding/_components/index.ts
@@ -1,6 +1,2 @@
-export { default as OnBoardingButton } from "./OnBoardingButton";
-export { default as OnBoardingCategories } from "./OnBoardingCategories";
-export { default as OnBoardingCompanies } from "./OnBoardingCompanies";
-export { default as OnBoardingMoods } from "./OnBoardingMoods";
-export { default as OnBoardingPriorities } from "./OnBoardingPriorities";
-export { default as OnBOardingTitle } from "./OnBoardingTitle";
+export { default as OnBoardingLoading } from "./onBoardingLoading";
+export { default as OnBoardingContainer } from "./onBoardingStep";
